backend/requirements.txt
# FastAPI and Server
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
email-validator==2.1.1

# Validation and Serialization
pydantic==2.6.3
pydantic[email]==2.6.3

# File Handling
python-magic==0.4.27
aiofiles==23.2.1

# CORS
starlette==0.36.3

# Testing
pytest==8.0.2
httpx==0.27.0